requirements.txt
Flask==3.1.0
gunicorn==21.2.0
Werkzeug>=3.1
psycopg2-binary==2.9.9
